Move Label test to its place
